requirements.txt
torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio==2.2.2+cpu
-f https://download.pytorch.org/whl/torch_stable.html
Flask
transformers
Pillow
langchain
chromadb
sentence-transformers
pandas
langchain-community
numpy<2

